requirements.txt
flask
flask-cors
sqlalchemy
pymysql
python-dotenv
joblib
scikit-learn==1.6.1
